Behavioral evals framework.
