+++ evals/logs/should_be_able_to_save_to_memory.log
@@ -1,0 +1,10 @@
+[
+  {
+    "toolRequest": {
+      "name": "save_memory",
+      "args": "{\n  \"fact\": \"My favorite color is blue\"\n}",
+      "success": true,
+      "duration_ms": 2
+    }
+  }
+]
