+++ evals/vitest.config.ts
@@ -1,0 +1,15 @@
+/**
+ * @license
+ * Copyright 2025 Google LLC
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  test: {
+    testTimeout: 300000, // 5 minutes
+    reporters: ['default'],
+    include: ['**/*.eval.ts'],
+  },
+});
