+++ packages/test-utils/package.json
@@ -9,6 +9,12 @@
     "build": "node ../../scripts/build_package.js",
     "typecheck": "tsc --noEmit"
   },
+  "dependencies": {
+    "@google/gemini-cli-core": "file:../core",
+    "@lydell/node-pty": "1.1.0",
+    "strip-ansi": "^7.1.2",
+    "vitest": "^3.2.4"
+  },
   "devDependencies": {
     "typescript": "^5.3.3"
   },
