+++ packages/test-utils/src/index.ts
@@ -5,3 +5,4 @@
  */
 
 export * from './file-system-test-helpers.js';
+export * from './test-rig.js';
